requirements.txt
requests
pandas
ta
schedule
python-dotenv
